rename CustomAccordionX to StyledAccordionX
