+++ packages/react-aria-widgets/src/Accordion/StyledAccordionHeader.tsx
