+++ packages/react-aria-widgets/src/Accordion/StyledAccordionPanel.tsx
